Rename some files. No need for ConnectionTask trait (removed).
